Fixes #3
I have updated line 844 to declare the Xml429 object and also
uncommented line 919 and confirmed that it will now build.

I have also updated the project settings to use the new ARINC 429 SDK
that we have just released this week (v4.6.0).

So, v4.6.0 will now have to be installed in order to build. v4.6.0 can
be downloaded from:
(32-bit)
http://www.aviftech.com/files/2213/8055/4804/sdk429-x86-4.6.0-win.zip

(64-bit)
http://www.aviftech.com/files/9113/8055/5530/sdk429-x64-4.6.0-win.zip